rename reducer folder as per Pascal case
